requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
python-dotenv
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
huggingface-hub>=0.20.0
transformers>=4.36.0
torch>=2.0.0
numpy>=1.24.0
